+++ Hearthstone Deck Tracker/Hearthstone/CounterSystem/Counters/FriendlyMinionsDiedThisGameCounter.cs
@@ -1,7 +1,9 @@
-﻿using HearthDb.Enums;
+﻿using System.Linq;
+using HearthDb.Enums;
 using Hearthstone_Deck_Tracker.Hearthstone.Entities;
 using Hearthstone_Deck_Tracker.LogReader.Interfaces;
 using Hearthstone_Deck_Tracker.Utility;
+using NuGet;
 
 namespace Hearthstone_Deck_Tracker.Hearthstone.CounterSystem.Counters;
 
@@ -25,14 +27,34 @@ public class FriendlyMinionsDiedThisGameCounter: NumericCounter
 		if(!Game.IsTraditionalHearthstoneMatch) return false;
 		if(IsPlayerCounter)
 			return InPlayerDeckOrKnown(RelatedCards);
-		return OpponentMayHaveRelevantCards() && Counter >= 10;
+
+		var couldBeGenerated = Game.Opponent.PlayerEntities.Any(e =>
+		{
+			if(e is { IsInHand: false, IsInPlay: false, IsInDeck: false })
+				return false;
+
+			if(!e.Info.Created)
+				return false;
+
+			if(e.HasCardId && e.CardId != HearthDb.CardIds.Collectible.Mage.Aessina)
+				return false;
+
+			var creatorId = e.Info.GetCreatorId();
+			if(creatorId <= 0 || !Game.Entities.TryGetValue(creatorId, out var creator))
+				return false;
+			return creator.CardId == HearthDb.CardIds.Collectible.Neutral.MalorneTheWaywatcher;
+		});
+
+		return (OpponentMayHaveRelevantCards() && Counter >= 10) || couldBeGenerated;
 	}
 
 	public override string[] GetCardsToDisplay()
 	{
-		return IsPlayerCounter ?
-			GetCardsInDeckOrKnown(RelatedCards).ToArray() :
-			FilterCardsByClassAndFormat(RelatedCards, Game.Opponent.OriginalClass);
+		if(IsPlayerCounter)
+			return GetCardsInDeckOrKnown(RelatedCards).ToArray();
+
+		var cardsFilteredByClassAndFormat = FilterCardsByClassAndFormat(RelatedCards, Game.Opponent.OriginalClass);
+		return cardsFilteredByClassAndFormat.IsEmpty() ? new[] { HearthDb.CardIds.Collectible.Mage.Aessina } : cardsFilteredByClassAndFormat;
 	}
 
 	public override string ValueToShow() => Counter.ToString();