+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## **Unreleased**
 **Hearthstone**:
+- Changed Aessina counter to show when opponent generated a Wild God from Malorne the Waywatcher.
 - Fixed an issue causing Oasis Ally to not appear in the secret tracker.
 
 ## **Release v1.43.3 - 2025-03-28**